requirements.txt
# Core dependencies
streamlit>=1.19.0,<2.0.0
streamlit-extras>=0.7.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0
openpyxl>=3.1.2
plotly>=5.15.0

# Database
sqlalchemy>=2.0.0

# Web scraping and automation
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.10.0
webdriver-manager>=4.0.0
playwright>=1.49.0
fake-useragent>=1.1.1

# AI models
google-generativeai>=0.3.0
pinecone>=5.0.0
litellm>=1.53.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
tqdm>=4.66.1
asyncio>=3.4.3
aiohttp>=3.8.5
tenacity>=8.2.3

# Optional: Crawl4AI for enhanced scraping capabilities
crawl4ai[all]>=0.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
